KES calculation update and prereqs fix
KES calculation moved from CNTools & gLiveView into a common function in env file. It will by default use node metrics with provided pool KES start period as fallback.

prereqs.sh script changed to properly handle non-existent branches and not try to replace content of script if file was never downloaded or if content is corrupt.

+++ scripts/cnode-helper-scripts/prereqs.sh
@@ -416,9 +416,9 @@ $sudo chown -R "$U_ID":"$G_ID" "${CNODE_HOME}" 2>/dev/null
 echo "Downloading files..."
 
 pushd "${CNODE_HOME}"/files >/dev/null || err_exit
-curl -s -f -m ${CURL_TIMEOUT} -o config.json.tmp ${URL_RAW}/files/config-mainnet.json 2>/dev/null
 
-if grep -i '404: Not Found' config.json.tmp >/dev/null ; then
+
+if ! curl -s -f -m ${CURL_TIMEOUT} -o config.json.tmp ${URL_RAW}/files/config-mainnet.json 2>/dev/null ; then
   err_exit "ERROR!! Specified branch could not be found! Kindly re-check the branch name and internet connection from the server"
 else
   echo "${BRANCH}" > "${CNODE_HOME}"/scripts/.env_branch
@@ -477,9 +477,18 @@ sed -e "s@/opt/cardano/cnode@${CNODE_HOME}@g" -e "s@CNODE_HOME@${CNODE_VNAME}_HO
 ### Update file retaining existing custom configs
 updateWithCustomConfig() {
   file=$1
+  if [[ ! -f ${file}.tmp ]]; then
+    echo "ERROR!! Failed to download '${file}' from GitHub"
+    return
+  fi
   if [[ -f ${file} && ${FORCE_OVERWRITE} = 'N' ]]; then
     if grep '^# Do NOT modify' ${file} >/dev/null 2>&1; then
       TEMPL_CMD=$(awk '/^# Do NOT modify/,0' ${file}.tmp)
+      if [[ -z ${TEMPL_CMD} ]]; then
+        echo "ERROR!! Script downloaded from GitHub corrupt, ignoring update for '${file}'"
+        rm -f ${file}.tmp
+        return
+      fi
       STATIC_CMD=$(awk '/#!/{x=1}/^# Do NOT modify/{exit} x' ${file})
       printf '%s\n%s\n' "${STATIC_CMD}" "${TEMPL_CMD}" > ${file}.tmp
     else
